task5/requirements.txt
flask==2.0.1
psycopg2-binary==2.9.9
werkzeug==2.0.3